test: add test helpers
