+++ src/utils/providers.ts
@@ -1,6 +1,6 @@
 import { providers } from "ethers"
 
-function getAlchemySubdomain(chainId: number): string | null {
+export function getAlchemySubdomain(chainId: number): string | null {
   switch (chainId) {
     case 1:
       return "eth"
@@ -19,15 +19,5 @@ export function buildAlchemyProvider(
 ): providers.JsonRpcProvider {
   const subdomain = getAlchemySubdomain(chainId)
   const url = `https://${subdomain}-mainnet.g.alchemy.com/v2/${apiKey}`
-  return new providers.JsonRpcProvider(url)
-}
-
-export function buildAlchemyProviderPredefined(
-  chainId: number,
-): providers.JsonRpcProvider {
-  const subdomain = getAlchemySubdomain(chainId)
-  /* eslint-disable  @typescript-eslint/no-non-null-assertion */
-  const url = `https://${subdomain}-mainnet.g.alchemy.com/v2/${process.env
-    .ALCHEMY_API_KEY!}`
   return new providers.JsonRpcProvider(url)
 }
